+++ test/libsolidity/syntaxTests/parsing/trailing_dot3.sol
@@ -1,0 +1,4 @@
+contract test {
+	uint a = 2.
+// ----
+// ParserError: (29-29): Expected identifier but got end of source