+++ Changelog.md
@@ -9,6 +9,7 @@ Breaking Changes:
  * General: New keywords: ``calldata``
  * General: ``continue`` in a ``do...while`` loop jumps to the condition (it used to jump to the loop body). Warning: this may silently change the semantics of existing code.
  * Type Checker: Disallow arithmetic operations for Boolean variables.
+ * Disallow trailing dots that are not followed by a number.
 
 Language Features:
  * General: Allow appending ``calldata`` keyword to types, to explicitly specify data location for arguments of external functions.